refactor: extracting cleanupPromise
